+++ Project.WebApi.Entities/Models/Audit.cs
@@ -2,7 +2,7 @@
 {
     public partial class Audit
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         public string? Operation { get; set; }
 
@@ -12,11 +12,11 @@
 
         public string? NewValue { get; set; }
 
-        public decimal? RecordId { get; set; }
+        public int? RecordId { get; set; }
 
         public DateTime? ChangeDate { get; set; }
 
-        public decimal? ChangedById { get; set; }
+        public int? ChangedById { get; set; }
 
         public virtual ICollection<AuditEntry> AuditEntries { get; set; } = new List<AuditEntry>();
     }
