+++ Project.WebApi.Entities/Models/Menu.cs
@@ -2,7 +2,7 @@
 {
     public partial class Menu
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; } = null!;
 
@@ -12,9 +12,9 @@
 
         public string? Route { get; set; }
 
-        public decimal? ParentNavigationId { get; set; }
+        public int? ParentNavigationId { get; set; }
 
-        public decimal? Order { get; set; }
+        public int? Order { get; set; }
 
         public bool? IsActive { get; set; }
 
@@ -24,15 +24,15 @@
 
         public DateTime? CreatedAt { get; set; }
 
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
-        public decimal? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
 
         public DateTime? DeletedAt { get; set; }
 
-        public decimal? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
 
         public virtual User? CreatedByNavigation { get; set; }
