+++ Project.WebApi.Entities/Models/Role.cs
@@ -2,13 +2,13 @@
 {
     public partial class Role
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
         public string Name { get; set; } = null!;
 
         public string? Description { get; set; }
 
-        public decimal? Order { get; set; }
+        public int? Order { get; set; }
 
         public bool? IsActive { get; set; }
 
@@ -16,15 +16,15 @@
 
         public DateTime? CreatedAt { get; set; }
 
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
-        public decimal? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
 
         public DateTime? DeletedAt { get; set; }
 
-        public decimal? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
 
         public virtual User? CreatedByNavigation { get; set; }
