+++ Project.WebApi.Entities/Models/UserRole.cs
@@ -2,11 +2,11 @@
 {
     public partial class UserRole
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
-        public decimal? UserId { get; set; }
+        public int? UserId { get; set; }
 
-        public decimal? RoleId { get; set; }
+        public int? RoleId { get; set; }
 
         public bool? IsActive { get; set; }
 
@@ -14,15 +14,15 @@
 
         public DateTime? CreatedAt { get; set; }
 
-        public decimal? CreatedBy { get; set; }
+        public int? CreatedBy { get; set; }
 
         public DateTime? UpdatedAt { get; set; }
 
-        public decimal? UpdatedBy { get; set; }
+        public int? UpdatedBy { get; set; }
 
         public DateTime? DeletedAt { get; set; }
 
-        public decimal? DeletedBy { get; set; }
+        public int? DeletedBy { get; set; }
 
         public virtual User? CreatedByNavigation { get; set; }
