+++ Project.WebApi.Entities/Models/UserToken.cs
@@ -2,9 +2,9 @@
 {
     public partial class UserToken
     {
-        public decimal Id { get; set; }
+        public int Id { get; set; }
 
-        public decimal UserId { get; set; }
+        public int UserId { get; set; }
 
         public string RefreshToken { get; set; } = null!;
