+++ UserManagementServices/ViewModels/Req_RoleVM.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public string? Description { get; set; }
-        public decimal? Order { get; set; }
+        public int? Order { get; set; }
         public RoleGrantVM Grants { get; set; }
         public bool? IsActive { get; set; }
         public List<Res_MenuVM> Menu { get; set; }
