+++ UserManagementServices/ViewModels/Res_RoleVM.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public decimal? Order { get; set; }
+        public int? Order { get; set; }
         public bool? IsActive { get; set; }
     }
 }
